sprint-4/step-1 - ordersFeedReducer
