+++ src/services/reducers/index.ts
@@ -2,6 +2,7 @@ import { combineReducers } from 'redux';
 
 import { IMainState, mainReducer } from './main';
 import { ingredientsReducer, IngredientsState } from './ingredients';
+import { IOrderState, ordersFeedReducer } from './orders-feed';
 
 export type { IMainState } from './main';
 
@@ -10,9 +11,11 @@ export type { IngredientsState } from './ingredients';
 export const rootReducer = combineReducers({
   main: mainReducer,
   ingredients: ingredientsReducer,
+  orders: ordersFeedReducer,
 });
 
 export interface IRootState {
   main: IMainState;
   ingredients: IngredientsState;
+  orders: IOrderState;
 }
