--- src/services/reducers/main/initialAppState.ts
@@ -1,31 +1,0 @@
-import { IMainState } from './IMainState';
-
-
-export const InitialAppState: IMainState = {
-  sum: 0,
-  // isIngredientsFailed: false,
-  // isIngredientsLoaded: false,
-  // ingredients: [],
-  ingredientAmountMap: {},
-  selectedBun: undefined,
-  selectedParts: [],
-  orderFeed: [],
-  orderTotal: 0,
-  orderToday: 0,
-  isOrderRequest: false,
-  isOrderFailed: false,
-  isOrderSuccess: false,
-  currentTabIndex: 0,
-  tabs: [
-    {
-      name: 'Булки',
-    },
-    {
-      name: 'Соусы',
-    },
-    {
-      name: 'Начинки',
-    },
-  ],
-  showedOrders: []
-};
