+++ src/services/reducers/orders-feed/index.ts
@@ -1,0 +1,2 @@
+export { ordersFeedReducer } from './ordersFeedReducer';
+export type { IOrderState } from './ordersFeedState';
